fix(instigator-tick-logs): allow `tickId` to be `BigInt` (#22122)
## Summary & Motivation
As the title.

## How I Tested These Changes
shadow dagit

+++ python_modules/dagster-graphql/dagster_graphql/schema/instigation.py
@@ -540,7 +540,7 @@ class GrapheneInstigationState(graphene.ObjectType):
     runsCount = graphene.NonNull(graphene.Int)
     tick = graphene.Field(
         graphene.NonNull(GrapheneInstigationTick),
-        tickId=graphene.NonNull(graphene.Int),
+        tickId=graphene.NonNull(graphene.BigInt),
     )
     ticks = graphene.Field(
         non_null_list(GrapheneInstigationTick),
